requirements.txt
Flask==3.0.3
pandas==2.2.3
openpyxl==3.1.5
fpdf2==2.7.9
python-barcode==0.15.1
Pillow==11.0.0
reportlab==4.2.5
